Another one bites the dust
sooo it didn't like " \t " or "\t". This is so weird, because it definitely works in my console tests, but not in the real thing! Although it makes me extremely nervous, I'm going to try separating things by actually hitting the tab key. Tab just doesn't look like a tab on this website! I have a strong feeling that it's where I've been going wrong.